Better code org and explanation
